requirements.txt
joblib==1.2.0
cherrypy==18.8.0
numpy==1.23.5
pandas==1.5.3
flask==2.2.3
python-dotenv==1.0.0
